выделяет логику карточки фильма в отдельный презентер
